PCP-55: new_pcp_map_response now uses a struct in6_addr instead of string
The input to the new_pcp_map_response function now uses an IPv6 struct
instead of a string as it probably should have. This will make it
easier to pass a parameter to save the mapping to apteryx and possibly
other future tasks. Also replaced a memcpy with direct assignment as it
does the same thing and is more readable.

Change-Id: Ia69579a1b2b829ccaee4fc2425468c5f1a623887
Reviewed-by: Tobi Wulff <tobi.wulff@alliedtelesis.co.nz>