PCP-55: New end_of_life variable for mapping struct
New member of the pcp mapping struct called end_of_life. This was
introduced so the start_of_life field can be left constant for a user
to look at when the mapping was started after the mapping's life has
been renewed. The remaining lifetime get function has also been
updated to use the new value along with an updated unit test.

Change-Id: Icd2826b7826a756187474063da3980f2a2c6a509
Reviewed-by: Tobi Wulff <tobi.wulff@alliedtelesis.co.nz>