PCP-57: Unit tests for mapping callbacks and removed a goto
New unit tests for the mapping callbacks. Mainly just counts how many times
the new/delete pcp mapping functions are called and compares counts of
how many mappings are and should be present. Removed the goto error in
the mapping find function as it was not needed. Tests for adding or
deleting duplicate mappings also included.

Change-Id: I7de67ae8bf687a7f58379afce9003690aa6986f1
Reviewed-by: Tobi Wulff <tobi.wulff@alliedtelesis.co.nz>