PCP-60: Duplicate mappings are detected and handled correctly
When a map request is received by the daemon it now performs checks to
whether a mapping with the same nonce, internal ip, internal port, and
protocol exists already. If it does, it bumps the end_of_life value
of that mapping both locally and in apteryx. Otherwise it behaves the
same as before.

Change-Id: I103add8be8a3e98bfa4c076f734127c27c944a5b
Reviewed-by: Tobi Wulff <tobi.wulff@alliedtelesis.co.nz>