PCP-64: Function clean up to prepare for adding validation checks
Functions were remade so validation checks could be more easily
added in. Debug printing was moved to its own temporary function. Macros
for separating r_opcode in headers were introduced. The MAP response
from processing the request is now created at the beginning and
constantly modified throughout the process function. New functions for
creating MAP responses and generic PCP responses. Added many TODOs as a
guide to what to add in the future to follow the RFC. New function
which attempts to create or extend a mapping given the map request and
response at that time. Lifetime refreshes now also refreshes the lifetime
value and not just the end of life.

Change-Id: I6a1a8315f888d7cfa405e70ef6e767d5a55ff063
Reviewed-by: Tobi Wulff <tobi.wulff@alliedtelesis.co.nz>