PCP-64: Validation of packets added and other improvements
Most of the packet validation as described on pages 24 and 25 along with
the necessary support functions have been added. Checks unsupported
versions, unsupported opcodes and packets that are too long or too short.
Support functions include getting the versions and r_opcode values
directly from the packet buffer, getting whether or not the R bit is set,
creating an error response, etting the lifetime to be the min or max if
the requested lifetime is too long or too short and getting error lifetimes
from a given result code. Moved process_map_request out of the run_loop
function to shorten its length and to make it easier to add more
process_x_request functions of other opcodes.

Change-Id: I77d163d9414e3a4f745698bd0ce75b9cb1aaaa5f
Reviewed-by: Tobi Wulff <tobi.wulff@alliedtelesis.co.nz>