PCP-66: Unit tests for refreshing lifetimes and detect zero lifetime
Added function comments for the refresh lifetime function and now it
checks if the given end_of_life value is sane. Setting the return value
in that function makes more sense. Receiving a lifetime of zero is now
detected and is used to delete mappings. New unit tests to test the
refresh lifetime function in both valid and invalid cases.

Change-Id: I501fec14377d111b904159e6002fce973df22fbc
Reviewed-by: Tobi Wulff <tobi.wulff@alliedtelesis.co.nz>