PCP-70: New startup time functionality and updated write functions
New functionality to track the startup time of the server so the uptime
can easily be calculated. TIME_BUF_SIZE moved to header file so pcpd
can use it too. Updated temporary debug print functions to verify that it
works. Mappings are now printed in the write state function with the same
output as in IMI's show command. Added a TODO for the static mappings
section of the write. Ctime removed as it was marked deprecated by
POSIX.1-2008 and replaced with strftime. New function for getting uptime
string since existing library functions for time strings assume input is
always a time from the epoch.

Change-Id: I4f668dbdd276cc66a9f8442af85d56b8daa85335
Reviewed-by: Tobi Wulff <tobi.wulff@alliedtelesis.co.nz>