PCP-72: Port mappings automatically created in iptables if valid request
Added new functionality to add bidirectional port mappings into iptables.
A new one is added each time a packet is successfully parsed and is a
valid request. Currently cannot delete mappings, does not support IPv6,
and error result codes are not received properly. Rearranged lines in
pcpd and a new function in pcp.c to sync the index numbers between the
ones stored locally/in apteryx and the chains in iptables.

Change-Id: Ia58d9db28fe90ea30b1b233c044b489499ef8025
Reviewed-by: Tobi Wulff <tobi.wulff@alliedtelesis.co.nz>