PCP-74: Moved delete chain to callback and iptables cleared when pcpd exits
Before, the clear pcp command would clear the mappings locally in the
daemon and in apteryx but not in iptables. Now the delete function was
moved to the callback function so iptables entries will be deleted as
apteryx entries are deleted. Rearranged function order and introduced for
loop when exiting pcpd so iptables entries for each mapping are manually
removed before exiting. Replaced int return of either 0 or -1 in the chain
write/remove functions to bools for consistency. Struct args in the
write chain function now uses pointers.

Change-Id: I497af6eebc91f10291fd7ce6e8660e2431b15d60
Reviewed-by: Tobi Wulff <tobi.wulff@alliedtelesis.co.nz>